applications/workbench/requirements.txt
cmake
mxnet
mpld3
bokeh
Pillow
ipykernel
ipython-genutils
ipython
ipywidgets
jupyter
jupyter-client
jupyter-contrib-nbextensions
jupyter-core
mapboxgl
matplotlib
numpy
opencv_python
pandas
python_magic
scikit-image
scikit-learn
scipy
umap-learn
tensorflow
wordcloud
holoviews
tqdm
MulticoreTSNE
